+++ src/vmobjects/VMArray.h
@@ -34,13 +34,13 @@ THE SOFTWARE.
 
 class VMArray : public VMObject {
 	public:
-		VMArray(int size, int nof=0);
+		VMArray(long size, long nof=0);
 #ifdef USE_TAGGING
 		virtual void WalkObjects(AbstractVMObject* (AbstractVMObject*));
 #else
 		virtual void WalkObjects(pVMObject (pVMObject));
 #endif
-		inline int GetNumberOfIndexableFields() const;
+		inline long GetNumberOfIndexableFields() const;
 		pVMArray    CopyAndExtendWith(pVMObject) const;
 		pVMObject   GetIndexableField(long idx) const;
 		void        SetIndexableField(long idx, pVMObject value);
@@ -52,12 +52,12 @@ class VMArray : public VMObject {
 #endif
 
 	private:
-		static const int VMArrayNumberOfFields;
+		static const long VMArrayNumberOfFields;
 };
 
-int VMArray::GetNumberOfIndexableFields() const {
+long VMArray::GetNumberOfIndexableFields() const {
 	static const void* cachedArray = NULL;
-	static int noOfIndexableFields = -1;
+	static long noOfIndexableFields = -1;
 
 	if(this != cachedArray) {
 		noOfIndexableFields = this->GetAdditionalSpaceConsumption() / sizeof(pVMObject);
